move re_args to utils.py
